+++ Modules/Models/Models/Entity.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 public protocol Entity {
-    var id: Int { get }
+    var id: Int64 { get }
 }
